+++ src/cpp/src/database/Database.test.cpp
@@ -13,12 +13,12 @@ namespace pairinteraction {
 DOCTEST_TEST_CASE("get a KetAtom") {
     Database &database = Database::get_global_instance();
 
-    AtomDescriptionByParameters<float> description;
+    AtomDescriptionByParameters<double> description;
     description.quantum_number_n = 60;
     description.quantum_number_l = 0;
     description.quantum_number_m = 0.5;
 
-    auto ket = database.get_ket<float>("Rb", description);
+    auto ket = database.get_ket<double>("Rb", description);
 
     DOCTEST_MESSAGE("KetAtom: ", *ket);
 }
@@ -26,45 +26,45 @@ DOCTEST_TEST_CASE("get a KetAtom") {
 DOCTEST_TEST_CASE("too large quantum number m") {
     Database &database = Database::get_global_instance();
 
-    AtomDescriptionByParameters<float> description;
+    AtomDescriptionByParameters<double> description;
     description.quantum_number_n = 60;
     description.quantum_number_l = 0;
     description.quantum_number_m = 1.5;
 
-    DOCTEST_CHECK_THROWS(database.get_ket<float>("Rb", description));
+    DOCTEST_CHECK_THROWS(database.get_ket<double>("Rb", description));
 }
 
 DOCTEST_TEST_CASE("not uniquely specified ket") {
     Database &database = Database::get_global_instance();
 
-    AtomDescriptionByParameters<float> description;
+    AtomDescriptionByParameters<double> description;
     description.quantum_number_n = 60;
     description.quantum_number_l = 0.9;
     description.quantum_number_m = 0.5;
 
-    DOCTEST_CHECK_THROWS(database.get_ket<float>("Rb", description));
+    DOCTEST_CHECK_THROWS(database.get_ket<double>("Rb", description));
 }
 
 DOCTEST_TEST_CASE("uniquely specified ket") {
     Database &database = Database::get_global_instance();
 
-    AtomDescriptionByParameters<float> description;
+    AtomDescriptionByParameters<double> description;
     description.quantum_number_n = 60;
     description.quantum_number_l = 0.9;
     description.quantum_number_j = 0.5;
     description.quantum_number_m = 0.5;
 
-    DOCTEST_CHECK_NOTHROW(database.get_ket<float>("Rb", description));
+    DOCTEST_CHECK_NOTHROW(database.get_ket<double>("Rb", description));
 }
 
 DOCTEST_TEST_CASE("get a BasisAtom") {
     Database &database = Database::get_global_instance();
 
-    AtomDescriptionByRanges<float> description;
+    AtomDescriptionByRanges<double> description;
     description.range_quantum_number_n = {60, 60};
     description.range_quantum_number_l = {0, 1};
 
-    auto basis = database.get_basis<float>("Rb", description, {});
+    auto basis = database.get_basis<double>("Rb", description, {});
 
     for (const auto &ket : *basis) {
         DOCTEST_MESSAGE("KetAtom: ", *ket);
@@ -74,14 +74,14 @@ DOCTEST_TEST_CASE("get a BasisAtom") {
 DOCTEST_TEST_CASE("get an OperatorAtom") {
     Database &database = Database::get_global_instance();
 
-    AtomDescriptionByRanges<float> description;
+    AtomDescriptionByRanges<double> description;
     description.range_quantum_number_n = {60, 60};
     description.range_quantum_number_l = {0, 1};
 
-    auto basis = database.get_basis<float>("Rb", description, {});
+    auto basis = database.get_basis<double>("Rb", description, {});
 
     auto dipole =
-        database.get_matrix_elements<float>(basis, basis, OperatorType::ELECTRIC_DIPOLE, 0);
+        database.get_matrix_elements<double>(basis, basis, OperatorType::ELECTRIC_DIPOLE, 0);
 
     DOCTEST_MESSAGE("Number of basis states: ", basis->get_number_of_states());
     DOCTEST_MESSAGE("Number of non-zero entries: ", dipole.nonZeros());
